+++ src/scripts/actionManager/faActionManagerHistory_manager.ts
@@ -1,9 +1,13 @@
-import { buildFaActionPayloadPreview } from './faActionManagerErrorReporting_manager'
+import {
+  buildFaActionErrorOrWarningPayloadPreview,
+  buildFaActionPayloadPreview
+} from './faActionManagerErrorReporting_manager'
 import { resolveFaActionManagerStore } from './faActionManagerStoreBridge_manager'
 
 import { createFaActionManagerHistory } from './functions/createFaActionManagerHistory'
 
 const faActionManagerHistoryApi = createFaActionManagerHistory({
+  buildFaActionErrorOrWarningPayloadPreview,
   buildFaActionPayloadPreview,
   resolveFaActionManagerStore
 })